desarrollo UserService

+++ src/main/java/cl/praxis/bootcapp/services/IBaseServiceCRUD.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface IBaseServiceCRUD<T> {
     List<T> getAll();
-    T getById(int id);
+    T getById(Long id);
     T create(T t);
     T update(T t);
-    boolean delete(int id);
+    boolean delete(Long id);
 }
